Address moderation PR review feedback
